requirements.txt
python-dotenv
playwright
bs4